+++ CHANGELOG.md
@@ -21,6 +21,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Don't reject actor if associated public key has different origin.
 - Set `Accept` header value to `application/jrd+json` when performing WebFinger query.
 
+### Fixed
+
+- Fixed WebFinger lookup not working on Yggdrasil addresses.
+
 ## [4.3.1] - 2025-05-18
 
 ### Fixed